+++ librespectrum/src/devs/mem/dynamic_48k.rs
@@ -51,7 +51,7 @@ impl Memory for Dynamic48k {
 }
 
 impl Identifiable for Dynamic48k {
-    fn id(&self) -> u32 { 2 }
+    fn id(&self) -> usize { 2 }
 }
 
 impl Device for Dynamic48k {
